+++ apps/docs/src/app/core/api-files.ts
@@ -38,7 +38,8 @@ export const API_FILES = {
     checkbox: ['CheckboxComponent'],
     datePicker: ['DatePickerComponent', 'DateFormatParser', 'DateFormatParserDefault', 'FdDate'],
     datetimePicker: ['DatetimePickerComponent', 'FdDateTime', 'DateTimeFormatParser', 'DateTimeFormatParserDefault'],
-    fileInput: ['FileInputComponent', 'FileSelectDirective', 'FileDragndropDirective'],
+    dropdown: ['PopoverDropdownComponent'],
+    fileInput: ['FileInputComponent', 'FileInputSelectDirective', 'FileInputDragndropDirective'],
     fileUploader: ['FileUploaderComponent', 'FileUploaderSelectDirective', 'FileUploaderDragndropDirective'],
     form: [
         'FormControlComponent',