+++ apps/docs/src/app/platform/api-files.ts
@@ -27,5 +27,7 @@ export const API_FILES = {
         'PlatformPanelConfig'
     ],
     switch: ['SwitchComponent'],
-    stepInput: ['NumberStepInputComponent', 'StepInputComponent', 'StepInputConfig']
+    stepInput: ['NumberStepInputComponent', 'StepInputComponent', 'StepInputConfig'],
+    inputGroup: ['InputGroupConfig', 'InputGroupComponent', 'InputGroupAddonComponent', 'InputGroupInputComponent'],
+    combobox: ['ComboboxComponent', 'BaseCombobox', 'ComboboxConfig']
 };