+++ apps/docs/src/app/platform/component-docs/platform-forms/input-group/platform-input-group-examples/platform-input-group-form-example.component.ts
@@ -1,0 +1,8 @@
+import { Component, ChangeDetectionStrategy } from '@angular/core';
+
+@Component({
+    selector: 'fdp-input-group-form-example',
+    templateUrl: './platform-input-group-form-example.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class PlatformInputGroupFormExampleComponent {}