+++ apps/docs/src/app/platform/platform-documentation.routes.ts
@@ -37,7 +37,7 @@ export const ROUTES: Routes = [
                 loadChildren: () =>
                     import(
                         './component-docs/platform-forms/platform-checkbox-group/platform-checkbox-group-docs.module'
-                    ).then((m) => m.PlatformCheckboxGroupDocsModule)
+                        ).then((m) => m.PlatformCheckboxGroupDocsModule)
             },
             {
                 path: 'link',
@@ -132,6 +132,19 @@ export const ROUTES: Routes = [
                     import('./component-docs/platform-object-status/platform-object-status-docs.module').then(
                         (m) => m.PlatformObjectStatusDocsModule
                     )
+            },
+            {
+                path: 'input-group',
+                loadChildren: () =>
+                    import('./component-docs/platform-forms/input-group/platform-input-group-docs.module').then(
+                        (m) => m.PlatformInputGroupDocsModule
+                    )
+            },
+            {
+                path: 'combobox',
+                loadChildren: () =>
+                    import('./component-docs/platform-combobox/platform-combobox-docs.module')
+                        .then(m => m.PlatformComboboxDocsModule)
             }
         ]
     }