+++ libs/platform/src/lib/domain/public_api.ts
@@ -1,3 +1,4 @@
 export * from './array-data-source';
+export * from './observable-data-source';
 export * from './data-model';
 export * from './data-source';