+++ libs/platform/src/lib/platform.module.ts
@@ -19,7 +19,9 @@ import { PlatformCheckboxGroupModule } from './components/form/checkbox-group/ch
 import { PlatformPanelModule } from './components/panel/panel.module';
 import { PlatformSwitchModule } from './components/form/switch/switch.module';
 import { PlatformObjectStatusModule } from './components/object-status/object-status.module';
-import { PlatformInputModule, PlatformStepInputModule } from './components/form/public_api';
+import { PlatformInputModule, PlatformStepInputModule, PlatformInputGroupModule } from './components/form/public_api';
+import { PlatformComboboxMobileModule } from './components/form/combobox/combobox-mobile/combobox-mobile.module';
+import { PlatformComboboxModule } from './components/form/combobox';
 
 @NgModule({
     imports: [CommonModule, FundamentalNgxCoreModule],
@@ -45,7 +47,10 @@ import { PlatformInputModule, PlatformStepInputModule } from './components/form/
         PlatformTextAreaModule,
         PlatformObjectStatusModule,
         PlatformInputModule,
-        PlatformStepInputModule
+        PlatformStepInputModule,
+        PlatformInputGroupModule,
+        PlatformComboboxModule,
+        PlatformComboboxMobileModule
     ]
 })
 export class FundamentalNgxPlatformModule {}